+++ Moth/LLVM/Metadata/Type.cs
@@ -7,8 +7,8 @@ public struct Type
     public bool is_foreign;
     public bool is_union;
     public PrivacyType privacy;
-    public ulong name_table_index;
-    public ulong name_table_length;
-    public ulong field_table_index;
-    public ulong field_table_length;
+    public uint name_table_index;
+    public uint name_table_length;
+    public uint field_table_index;
+    public uint field_table_length;
 }
